内部の音声処理をFloatに変更 クリップランプを実装

+++ src/main/kotlin/objects/AudioSampleProvider.kt
@@ -1,5 +1,5 @@
 package objects
 
 interface AudioSampleProvider {
-    fun getSamples(frame : Int):ShortArray
+    fun getSamples(frame : Int):FloatArray
 }
